# Perf: Prevent unnecessary page decoding